Moves shortcodes to .eleventy directory
