+++ .eleventyignore
@@ -6,3 +6,4 @@ package.json
 README.md
 Vagrantfile
 _site/
+.eleventy/
